Address most warnings

+++ include/display/effects/utilities.h
@@ -16,7 +16,7 @@ using Generator = std::function<flm::CRGB()>;
 
 // Colour generating functions
 inline Generator randomHSV = []() { return flm::CHSV(flm::random8(), 255, 255); };
-inline Generator cycleHSV = []() { return flm::CHSV((millis() / 10), 255, 255); };
+inline Generator cycleHSV = []() { return flm::CHSV(static_cast<uint8_t>(millis() / 10), 255, 255); };
 inline Generator black = []() { return 0; };
 inline Generator white = []() { return flm::CRGB::White; };
